requirements.txt
django

django-extensions

django-environ

python-binance

openpyxl

pandas
